requirements.txt
asgiref==3.5.2
attrs==21.4.0
autopep8==1.6.0
certifi==2022.6.15
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
Django==4.0
djangorestframework==3.13.1
drf-spectacular==0.22.1
drf-yasg==1.20.0
flake8==4.0.1
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.6.1
MarkupSafe==2.1.1
mccabe==0.6.1
packaging==21.3
pycodestyle==2.8.0
pyflakes==2.4.0
pyparsing==3.0.9
pyrsistent==0.18.1
pytz==2022.1
PyYAML==6.0
requests==2.28.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
sqlparse==0.4.2
toml==0.10.2
uritemplate==4.1.1
urllib3==1.26.9
